fix: clicking theme button does not change colorcheme
This happens because the default theme is system but the ThemeChanger is
a button switching between the 'light' and 'dark' theme so in the first
run it would switch from 'system' to 'light' or 'dark' and might not
change the actual colorcheme if the system's one aligns to the one the
switcher would switch. Making the default 'light' solves that problem
while not making the switcher complicated in order to offer 3
possibilities.