feat: juans changes (#1)
* fix: clicking theme button does not change colorScheme

This happens because the default theme is system but the ThemeChanger is
a button switching between the 'light' and 'dark' theme so in the first
run it would switch from 'system' to 'light' or 'dark' and might not
change the actual colorcheme if the system's one aligns to the one the
switcher would switch. Making the default 'light' solves that problem
while not making the switcher complicated in order to offer 3
possibilities.

* fix: infered types in filter functions

* fix: hideTracker actually toggles

* fix?: remove funciton declaration inside function declaration -> use arrow func

this is more a vibe rather than a hard requirement lol. Also, JS is
weird, I'm sure that the arrow funcitons will be recreated every time the
component using the hook rerenders but the function declarations are
hoisted so I am not really sure if the same applies.

* chore: prettier